// src/metadatas.d.ts
export interface LambdaExpressionMetadata {
    expressionLeft: string;
    operator: string;
    expressionRight: string;
}
export interface LambdaPropertiesMetadata {
    propertiesLeft: string[];
    operator: string;
    propertiesRight: string[];
}
export interface LambdaColumnMetadata {
    columnLeft: string;
    operator: string;
    columnRight: string;
}